It should be Quickplay, also it had different effects at different times
My Proposed effect: 
Select 1 monter on your side of the field with less than 500 ATK and DEF to activate 1 of the following effects (if the monster is face-down, flip it face-up): ● If the monster is "Kuriboh," tribute it to Special Summon as many "Kuriboh Token" (s) as possible to your side of the field in face-up Defense Position. This card then remains face-up on the field. As long as this card is face-up on the field you can Special Summon as many "Kuriboh Token" (s) as you want this way to your side of the field during either player's turn. These tokens cannot be tributed for a Tribute Summon. ● When your opponent targets a monster you control: change the target to the selected monster, then destroy the monster and take 300 damage. ● Special Summon 2 "Copy Token" (s) to your side of the field with the exact same name, effect, Level, ATK, and DEF in the same position as the targetted monster. These tokens cannot declare an attack. These tokens can be tributed for a Tribute Summon. 

This covers every way that the anime effect was used in the show. In 22-24 and 71-73, he uses the first effect. In 38-39, he uses the 2nd, and in 139 he uses the third.