This effect did more than negate traps
Negate the effects of all other Trap Cards during the turn you activate this card. Then, select one monster on your opponent's side of the field. As long as this card remains face-up on the field, that monster cannot attack, change its battle position, or be targeted for an attack.

This is from 61, and the card should be a Continuous Trap instead of a Normal one.