A way to make this effect an equip effect like in the show
Proposed effect: 
Special Summon 1 "Parasite Paracide" from your opponent's Deck to their side of the field in face-up Attack Position. All monsters on your opponent's side of the field and those Summoned afterwards are equipped with a "Parasite Token" that is treated as an Equip Spell Card. Monsters equipped with a "Parasite Token" are treated as Insect-Type and cannot be Tributed. This effect lasts as long as "Parasite Paracide" and "Reckless Parasite" are on the field. 

(From 63 and 64) One thing is that right now the coding doesn't even match the effect we already have because Parasite Paracide gets summoned to your field instead of your opponent's. But besides that, if this effect worked this way it would make monsters like Gearfried immune like they were in the show. I get this is probably difficult (maybe impossible) to code tokens that can equip, but if so that would work a lot better. Also, this mirrors the official Wiki text: http://yugioh.wikia.com/wiki/Reckless_Parasite